==> 1.Input data from large number of text files.ipynb <==
import os
import pandas as pd
import numpy as np

data_folder1=os.path.join("D:\GF\input_2020",'1_5000')
data_folder2=os.path.join("D:\GF\input_2020",'5001_10000')
data_folder3=os.path.join("D:\GF\input_2020",'10001_end')
#importing data from local machine

data_folder1         #checking the path

files1=[]
data1=[]
for root, folders,files in os.walk(data_folder1):
    for file in files:
        path=os.path.join(root,file)
        files1.append(file.rstrip(".txt"))
        with open(path,"r",encoding="latin-1") as inf:
            data1.append(inf.read())                   #saving the file names into files1 and data of those files into data1 respectively

files2=[]
data2=[]
for root, folders,files in os.walk(data_folder2):
    for file in files:
        path=os.path.join(root,file)
        files2.append(file.rstrip(".txt"))
        with open(path,"r",encoding="latin-1") as inf:
            data2.append(inf.read())

files3=[]
data3=[]
for root, folders,files in os.walk(data_folder3):
    for file in files:
        path=os.path.join(root,file)
        files3.append(file.rstrip(".txt"))
        with open(path,"r",encoding="latin-1") as inf:
            data3.append(inf.read())

df_1_5000=pd.DataFrame({"File_name":files1,"Data":data1})
df_1_5000.head()

df_5001_10000=pd.DataFrame({"File_name":files2,"Data":data2})
df_5001_10000.head()

df_10001_end=pd.DataFrame({"File_name":files3,"Data":data3})
df_10001_end.head()

df_1_5000.shape,df_5001_10000.shape,df_10001_end.shape

Final_df=pd.concat([df_1_5000,df_5001_10000,df_10001_end],axis=0)
Final_df.head()                                                   #concatenating all the dataframes to form one single DataFrame

Final_df.to_csv("Complete_GF_data_2020.csv",index=False)  #saving whole DataFrame as csv file

Final_df.shape
