==> 2.Extracting data from given output excel file.ipynb <==
import pandas as pd
import numpy as np

xls = pd.ExcelFile(r'D:\Gianfranco SICILIANO\reshape_JanMay2020.xlsx')

# Now you can list all sheets in the file
sheet_names=list(xls.sheet_names)
sheet_names

dfex_new=pd.DataFrame()
for i in sheet_names[1:]:
    dfex_new[i]=pd.Series([j.title() for j in pd.read_excel(r'D:\Gianfranco SICILIANO\reshape_JanMay2020.xlsx', sheet_name=i).columns])

dfex_new.head(9)

dfex_new.iloc[8:,:].to_csv("GF_clean_words2020.csv", index=False)  #saving all the words which starts from 8th row

import pandas as pd
dfn=pd.read_excel(r'D:\Gianfranco SICILIANO\reshape_JanMay2020.xlsx', sheet_name="General info from parsing")
dfn.head()

dfn.shape

dfn=dfGP[["isin","disclosure"]]
dfn.head()

dfn["File_name_output"]=dfn["isin"].str.strip()+"_"+dfn["disclosure"].str.strip()

dfn.head()

dfn["File_name"]=dfn["File_name_output"].str.strip()

dfn.head()

dfn.drop(["words#"],axis=1,inplace=True)  
dfn.drop(["File_name_output"],axis=1,inplace=True)  

dfn.head()

dfn.shape

dfn.to_csv("File_name_output2020.csv",index=False)

dfl=pd.read_csv("Complete_GF_data_2020.csv")
dfl.head()

dfl1=pd.concat([dfl["File_name"].str.strip(),dfl["Data"]],axis=1)
dfl1.head()

dfl1.shape

dfl1[dfl1["File_name"]=="AT0000609607_pre"]

dfl.shape

dfl.shape,dfn.shape

File_name_output3=[]
File_name3=[]
Data3=[]
for i in dfn["File_name_output"].str.strip():
    for j in dfl["File_name"].str.strip():
        if i==j:
            File_name_output3.append(i)
            File_name3.append(j)
            Data3.append(dfl[dfl["File_name"].str.strip()==i]["Data"].values[0])             #code for merging input and output including duplicate values
df0kld = pd.DataFrame({"File_name": File_name3,"File_name_output":File_name_output3,
                    "Data": Data3})
df0kld.shape

df0kld.head()

Final=df0kld[["File_name","Data"]]
Final.shape

Final.to_csv("Final_GF_12726_2020.csv", index=False)

df0kld["File_name_output"].value_counts()       #No duplicate values are present

print("Unique: ",len(dfn["File_name_output"].unique()),",","Total: ",len(dfn["File_name_output"])) #Expected

print("Unique: ",len(df0kld["File_name_output"].unique()),",","Total: ",len(df0kld["File_name_output"])) #Actual

print("Unique: ",len(dfl["File_name"].unique()),",","Total: ",len(dfl["File_name"])) #Total

len(set(dfl["File_name"].str.strip())&set(dfn["File_name_output"].str.strip()))    #Common values

len(dfl["File_name"].str.strip().unique())   #No duplicates in original data

15807-12813,12813-12726

Rows_3081=list(set(dfl["File_name"].str.strip().unique())-set(dfn["File_name_output"].str.strip().unique()))
len(Rows_3081)                      #Total-output

Rows_3081[:7]   #total should not be in output

Rows_87a=list(set(dfn["File_name"].unique())-set(dfl1["File_name"].unique()))
len(Rows_87a)

pd.Series(Rows_87a).to_csv("No_data_87_2020.csv",index=False)

Rows_87=list(set(dfn["File_name_output"].str.strip().unique())-set(dfl["File_name"].str.strip().unique()))
len(Rows_87)   #output-total

Rows_87[:8]  #output not in total

dfl[dfl["File_name"].str.strip()=="GB00BZ58DH10_other1"] #output not in total

dfl[dfl["File_name"].str.strip()=="GB0002570660_other14"]

dfn[dfn["File_name_output"].str.strip()=="GB0002570660_other14"] #total not in output

dfn[dfn["File_name_output"].str.strip()=="FR0004125920_earn"]#total not in output

Conclusion: Unique Total input files-Output: 15807-12813=2994 Matching Total&Output: 12726, No data : 12813-12726=87(Need to get from client) Not present in output file: 3081(Perhaps analysis is not necessary)


New_missing_data_87=list(set(dfn["File_name_output"].str.strip().unique())-set(df0kld["File_name_output"].str.strip().unique()))
New_missing_data_87[:5]

len(New_missing_data_87)

Rows=set(Rows_3081)&set(New_missing_data_87)
len(Rows)                                                #No common in both

dfL=pd.read_excel(r'D:\GF\reshape_JanMay2020.xlsx', sheet_name="Litigious")
dfL.head()

dfL.shape
