==> 5.Counting of negative, positive words, etc.,.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os

%matplotlib inline
from urllib.request import urlopen
import nltk
from nltk.tokenize import word_tokenize  
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.stem.snowball import SnowballStemmer
from nltk.stem.porter import *
import spacy
import nltk
from nltk.tokenize.toktok import ToktokTokenizer
import re
import unicodedata

df1=pd.read_csv("Final_GF_data_2020.csv")
df1.head() 

df1.shape

df1["Data"].fillna('', inplace=True)

df1.shape

df2=pd.read_csv("GF_clean_words2020.csv")
df2.head(2)

DF1=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Negative"]} for j in range(len(df1["Data"]))])
DF1.to_csv("DF1_2020.csv",index=False)

df1["Data"][0].title().count("Press")

df1["Data"][0]

df1["Data"].dtype

DF1=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Negative"]} for j in range(len(df1["Data"]))])
DF1.to_csv("DF1_2020.csv",index=False)

DF1.head()

DF1.shape

a1=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF1],axis=1)
a1.head()

a1.shape

DF2=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Positive"]} for j in range(len(df1["Data"]))])
DF2.to_csv("DF2_2020.csv",index=False)

DF3=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Uncertainity"]} for j in range(len(df1["Data"]))])
DF3.to_csv("DF3_2020.csv",index=False)

DF4=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Litigious"]} for j in range(len(df1["Data"]))])
DF4.to_csv("DF4_2020.csv",index=False)

DF5=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Strong Modals"]} for j in range(len(df1["Data"]))])
DF5.to_csv("DF5_2020.csv",index=False)

a5=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF5],axis=1)
a5.head()

a5.shape

DF6=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Weak Modals"]} for j in range(len(df1["Data"]))])
DF6.to_csv("DF6_2020.csv",index=False)

DF7=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Risk"]} for j in range(len(df1["Data"]))])
DF7.to_csv("DF7_2020.csv",index=False)

DF8=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Risk Financial"]} for j in range(len(df1["Data"]))])
DF8.to_csv("DF8_2020.csv",index=False)

DF9=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Risk Other Idiosyncratic"]} for j in range(len(df1["Data"]))])
DF9.to_csv("DF9_2020.csv",index=False)

DF10=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Risk Legal and Regulatory"]} for j in range(len(df1["Data"]))])
DF10.to_csv("DF10_2020.csv",index=False)

DF11=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Risk Other Systematic"]} for j in range(len(df1["Data"]))])
DF11.to_csv("DF11_2020.csv",index=False)

DF12=pd.DataFrame([{str(i):df1["Data"][j].title().count(str(i)) for i in df2["Covid"]} for j in range(len(df1["Data"]))])
DF12.to_csv("DF12_2020.csv",index=False)

df1.shape

DF1=pd.read_csv("DF1_2020.csv")
DF2=pd.read_csv("DF2_2020.csv")
DF3=pd.read_csv("DF3_2020.csv")
DF4=pd.read_csv("DF4_2020.csv")
DF5=pd.read_csv("DF5_2020.csv")
DF6=pd.read_csv("DF6_2020.csv")
DF7=pd.read_csv("DF7_2020.csv")
DF8=pd.read_csv("DF8_2020.csv")
DF9=pd.read_csv("DF9_2020.csv")
DF10=pd.read_csv("DF10_2020.csv")
DF11=pd.read_csv("DF11_2020.csv")
DF12=pd.read_csv("DF12_2020.csv")

a0=pd.read_csv("General_info_from_parsing_2020.csv")
a0.head()

a1=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF1],axis=1)
a2=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF2],axis=1)
a3=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF3],axis=1)
a4=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF4],axis=1)
a5=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF5],axis=1)
a6=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF6],axis=1)
a7=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF7],axis=1)
a8=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF8],axis=1)
a9=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF9],axis=1)
a10=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF10],axis=1)
a11=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF11],axis=1)
a12=pd.concat([df1.drop(["Data"],axis=1,inplace=False),DF12],axis=1)

a1.to_csv("data_frame1_2020.csv",index=False)
a2.to_csv("data_frame2_2020.csv",index=False)
a3.to_csv("data_frame3_2020.csv",index=False)
a4.to_csv("data_frame4_2020.csv",index=False)
a5.to_csv("data_frame5_2020.csv",index=False)
a6.to_csv("data_frame6_2020.csv",index=False)
a7.to_csv("data_frame7_2020.csv",index=False)
a8.to_csv("data_frame8_2020.csv",index=False)
a9.to_csv("data_frame9_2020.csv",index=False)
a10.to_csv("data_frame10_2020.csv",index=False)
a11.to_csv("data_frame11_2020.csv",index=False)
a12.to_csv("data_frame12_2020.csv",index=False)

print(a0.shape,a1.shape,a2.shape,a3.shape,a4.shape,a5.shape,a6.shape,a7.shape,a8.shape,a9.shape,a10.shape,a11.shape,a12.shape)

with pd.ExcelWriter("D:\Gianfranco SICILIANO\GF_2020_Fin_Output.xlsx") as writer:
   
    # use to_excel function and specify the sheet_name and index
    # to store the dataframe in specified sheet
    a0.to_excel(writer, sheet_name="General_info_from_parsing", index=False)
    a1.to_excel(writer, sheet_name="Negative", index=False)
    a2.to_excel(writer, sheet_name="Positive", index=False)
    a3.to_excel(writer, sheet_name="Uncertainity", index=False)
    a4.to_excel(writer, sheet_name="Litigious", index=False)
    a5.to_excel(writer, sheet_name="Strong Modals", index=False)
    a6.to_excel(writer, sheet_name="Weak Modals", index=False)
    a7.to_excel(writer, sheet_name="Risk", index=False)
    a12.to_excel(writer, sheet_name="Covid", index=False)
    a8.to_excel(writer, sheet_name="Risk Financial", index=False)
    a9.to_excel(writer, sheet_name="Risk Other Idiosyncratic", index=False)
    a10.to_excel(writer, sheet_name="Risk Legal and Regulatory", index=False)
    a11.to_excel(writer, sheet_name="Risk Other Systematic", index=False)

# Converting float NAN in string column to string datatype

import pandas as pd
import numpy as np

#create DataFrame with some NaN values
dfz = pd.DataFrame({'team': ['A', 'A', 'A', 'A', 'B', 'B', 'B', 'B'],
                   'points': [np.nan, 11, 7, 7, 8, 6, 14, 15],
                   'assists': [5, np.nan, 7, 9, 12, 9, 9, 4],
                   'rebounds': [11, 8, 10, np.nan, 6, 5, 9, np.nan]})


dfz

dfz.isnull().sum()

dfz.to_csv("Nan_example_without_filling.csv",index=False)

df

df.isnull().sum()

df.fillna('', inplace=True)

df

df.isnull().sum()

df.to_csv("Nan_example.csv",index=False)
